### 2020-10-12 #2740 @luizirber
> @luizirber welcome to JOSS! Would you be able to edit this submission? Can't tell if this is close to your expertise at all.

It isn't my expertise, but I'm willing to try =]
