### 2020-10-25 c7d98d7
fix static build on Windows
