### 2020-10-25 ee754d4
hide symbols for GNU compilers
