### 2020-10-25 e77da04
Fixing SWIG with attribute
