### 2020-10-25 e9e94ba
Moved GCC attribute to ezc3d.i file
