### 2020-11-24 #2740 @Kevin-Mattheus-Moerman
@modenaxe @BKillen05 @melund @ixjlyons would you be interested in reviewing this work for [JOSS](https://joss.theoj.org/)?
The review focuses on [the software](https://github.com/pyomeca/ezc3d) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.02740/joss.02740/10.21105.joss.02740.pdf).
For more information have a look at our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
