### 2020-12-01 #2740 @melund
> @melund @pariterre says you made what sounds like a minor contribution to this project in the past. Can you confirm you do not have a conflict of interest reviewing this work? Thanks.

Yes. I can confirm that I don't have any conflicts of interests :)
