### 2020-12-02 #2740 @BKillen05
HI Kevin, Sorry for the radio silence I was finalising a fellowship application and let most of my mails lapse.  I am in the same boat as Luca at this point where I can likely find more time in January CheersBryce  Bryce A Killen, BExSc Hons PhDPostdoctoral Research FellowHuman Movement Biomechanics Research GroupKU Leuven, Leuven, BelgiumTwitter: @bryce_killen From: Kevin Mattheus MoermanSent: dinsdag 1 december 2020 12:34To: openjournals/joss-reviewsCc: BKillen05; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: ezc3d: An easy C3D file I/O cross-platform solution for C++, Python and MATLAB (#2740) @modenaxe If you are able to help in January that is fine too. I expect any reviewer will tell me "dude the holidays are coming", so I we can certainly wait if you are still interested.—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
