### 2020-12-07 2f77261
Prepare the analog for writing in the parameters file
