### 2020-12-12 #2740 @Kevin-Mattheus-Moerman
@BKillen05 @modenaxe since December/January is a holiday season for many we've seen this type of response from many editors/reviewers invited. This is fine and fully understandable and we can be flexible regarding timing to work with your constraints.

Would you be okay with being assigned as reviewers for now and to pick this up whenever you are "back in the office"? No worries if you'd rather not.
Thanks.
