### 2020-12-14 3820c97
Made things easier to create new points in Python
