### 2020-12-15 2326f7d
Added the capability to remove a parameter and a group
