### 2020-12-15 18a5cb8
Added a future formatting changer for Nexus
