### 2020-12-15 #2740 @BKillen05
Hi Kevin, That is fine for me CheersBryce  Bryce A Killen, BExSc Hons PhDPostdoctoral Research FellowHuman Movement Biomechanics Research GroupKU Leuven, Leuven, BelgiumTwitter: @bryce_killen From: Kevin Mattheus MoermanSent: zaterdag 12 december 2020 10:34To: openjournals/joss-reviewsCc: BKillen05; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: ezc3d: An easy C3D file I/O cross-platform solution for C++, Python and MATLAB (#2740) @BKillen05 @modenaxe since December/January is a holiday season for many we've seen this type of response from many editors/reviewers invited. This is fine and fully understandable and we can be flexible regarding timing to work with your constraints.Would you be okay with being assigned as reviewers for now and to pick this up whenever you are "back in the office"? No worries if you'd rather not.Thanks.—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
