### 2020-12-15 24ccd6d
Bumped the version in CMakeList
