### 2020-12-16 #2911 @whedon
[REVIEW]: ezc3d: An easy C3D file I/O cross-platform solution for C++, Python and MATLAB
**Submitting author:** @pariterre (<a href="http://orcid.org/0000-0002-5031-1048">Benjamin Michaud</a>)
**Repository:** <a href="https://github.com/pyomeca/ezc3d" target ="_blank">https://github.com/pyomeca/ezc3d</a>
**Version:** Release_1.4.2
