### 2021-01-05 #2911 @melund
Thanks a lot @Kevin-Mattheus-Moerman. But I still can't edit the checklist. And clicking the invitation link now gives me this:

![image](https://user-images.githubusercontent.com/1038978/103616397-acbaa300-4f2c-11eb-913b-62e99106b4ba.png)
