### 2021-01-07 551d320
Make offset always positive when reading analog data
