### 2021-01-07 4560a72
Bumped to version 1.3.9
