### 2021-01-15 pyomeca/ezc3d#177 @pariterre
Hi @Alan963852741
Could you please expand a bit? What OS are you using? Which Matlab version are you using? How did you compile? How and where did you install your the compiled version? Which folder did you "addpath" to?

Please note that ezc3d is not a Matlab library, it is a C++ library with a Matlab binder, so it needs to be compiled beforehand.
