### 2021-01-16 pyomeca/ezc3d#177 @Alan963852741
Thanks for your reply. The OS I used is Win10, and the version of Matlab is 2020b. I just download the code, of EZC3D and add the path of the code into Matlab. But it doesn't work when I use the command "ezc3dRead()".It comes to these problems.

“Function or variable'ezc3dRead' is not recognized."

So could you please tell me how to make it work? How to compile the code?
Thanks a lot.
