### 2021-01-16 pyomeca/ezc3d#177 @Alan963852741
> Hi @Alan963852741
> Could you please expand a bit? What OS are you using? Which Matlab version are you using? How did you compile? How and where did you install your the compiled version? Which folder did you "addpath" to?
>
> Please note that ezc3d is not a Matlab library, it is a C++ library with a Matlab binder, so it needs to be compiled beforehand.

Thanks for your reply. The OS I used is Win10, and the version of Matlab is 2020b. I just download the code, of EZC3D and add the path of the code into Matlab. But it doesn't work when I use the command "ezc3dRead()".It comes to these problems.

“Function or variable'ezc3dRead' is not recognized."

So could you please tell me how to make it work? How to compile the code?
Thanks a lot.
