### 2021-01-18 pyomeca/ezc3d#177 @pariterre
Hi @Alan963852741
As stated before, ezc3d is a C++ based library. This means that the core code is not Matlab, but C++. C++ codes cannot be run as is, but must be compiled. I have made a video in the past how to graphically compile C++ code.
You will need to install CMake and Visual Studio to compile
https://www.youtube.com/watch?v=gWno_NXrITA&feature=youtu.be
