### 2021-02-01 #2911 @melund
@Kevin-Mattheus-Moerman, I have finished my review and I am happy to accept the paper. The paper and the project fulfill all requirements for a JOSS publication.

The ezc3d is a valuable contribution which is already used in many projects. I believe it will receive many citations in the future.
I haven’t been able to assess the functionality of the C++ API but I have used the Python interface extensively which is just wrapper on the C++ API. Similarly I haven’t been able to try the Matlab API, but the documentation seems solid.

I have also contributed a few bug reports and fixes to the project prior to this review which were all handle well. I believe ezc3d is a best practice example of an open source project. Good work @pariterre
