### 2021-02-04 #2911 @BKillen05
@Kevin-Mattheus-Moerman Sorry for the delay in this review.
I have now read the paper and documentation and completed some testing for functionality etc.

I am extremely impressed by the quality of the submisison - both the software - descriptions - and documentation is extremely thorough. I am therefore more than happy to accept the paper and project in its current form. Similar to @melund  I have only tested the Python interface to date - if required I can try to carve out some time to try and do further testing on the MATLAB side.

I was previously able to edit the "checklist" but for some reason cannot do so now - I may need a new invitation ?
