### 2021-02-08 57309da
Add VCPKG section to README.md (fix #161)
