### 2021-02-10 pyomeca/ezc3d#177 @modenaxe
Hi @pariterre,
I follow up here as part of the JOSS reviewing process. It looks like the CMake file has evolved since the opening of the issue and it the default installation directory for the MATLAB files is now `...../MATLAB/R2020a/toolbox`.
Few issues:
1. building INSTALL on Visual Studio required me to run it as administrator in order to copy the files. This is mentioned in the CMake file but not in the instructions in the README.md and I think it should.
2. the installation copied the mex files into the `toolbox` folder but not the dll library, which I had to copy manually in order for ezc3d to work in MATLAB. Is that the supposed behaviour?
3. the MATLAB section of the documentation in the README.md should include the additional steps shown in the Youtube video, i.e. adding the toolbox to the path. Alternatively a setup utility script like the one used by OpenSim could be provided to help the users.
Thank you.
