### 2021-02-10 #2911 @modenaxe
hi @pariterre, I have [posted some questions](https://github.com/pyomeca/ezc3d/issues) under the `[JOSS review]` header.
