+++ README.md
@@ -13,9 +13,10 @@ So, without further ado, let's begin C3Ding!
 # Table of Contents  
 [How to install](#how-to-install)  
 [How to use](#how-to-use)  
-- [The C++ API](#the-c++-api)
+- [The C++ API](#the-c-api)
 - [MATLAB](#matlab)
-- [Python 3](#python-3)
+- [Python 3](#python-3)  
+
 [How to contribute](#how-to-contribute)  
 [Supported generated C3D](#supported-generated-c3d)  
 [Documentation](#documentation)  
