### 2021-02-10 56f2315
added direct link to documentation in readme
