### 2021-02-10 #2911 @modenaxe
@pariterre there are a couple more [here](https://github.com/pyomeca/ezc3d/issues/177) I added them to an existing Issue.
