### 2021-02-10 cdc5e1f
Added small description for vcpkg
