### 2021-02-10 pyomeca/ezc3d#177 @pariterre
@modenaxe

> building INSTALL on Visual Studio required me to run it as administrator in order to copy the files. This is mentioned in the CMake file but not in the instructions in the README.md and I think it should.

It was actually already written as such, but in a highly confusing manner. I rewrote the `MATLAB_ezc3d_INSTALL_DIR` description and tried to emphasized that point

> the installation copied the mex files into the toolbox folder but not the dll library, which I had to copy manually in order for ezc3d to work in MATLAB. Is that the supposed behaviour?

I'll have a look a this, this seems to be a proper bug... Since I do not use Windows nor Matlab and cannot travis test it, it probably have slipped... I have a look in the upcoming days!

> the MATLAB section of the documentation in the README.md should include the additional steps shown in the Youtube video, i.e. adding the toolbox to the path.

I am not sure where I could add this (or actually if I really want to, see below). The explanation for the `MATLAB_ezc3d_INSTALL_DIR` are already in the CMake section of `Compiling`. Do you mean in the Matlab's `How to use` section?

Still, I do not feel that including in the EZC3D documentation a documentation for CMake (other than the specificity of ezc3d) is something I actually can afford to do. CMake is rapidly changing, they develop like crazy over there! Offering any sort of support outside of their official documentation, is likely to be rapidly out-of-date and will become a huge and unwanted burden on me. I actively support EZC3D, RBDL-casadi, biorbd, bioviz and bioptim. For most of them, I am alone, and for all of them simply not paid.

I have kind of done it with the video that I added in the documentation, since I understand how it may help. However, I fear that is kind of a Pandora box for me. For instance, this video is highly specific to Windows users (which I don't even own since I run on Linux) with the CMake-GUI. But even the CMake GUI is not at all mandatory. For example, on Ubuntu, this GUI is not even available from the main repositories... The additional steps are not actual additional steps when you are using CMake not from the GUI, it is just part of the options (as described in CMake section of the `Compiling` section). Though I understand that CMake can be confusing and hard to use for a new programmer... I just feel that I do not have the time nor manpower to properly support a CMake documentation... And a bad documentation is worse than none I think...

I feel that my answer may be unsatisfactory... If it is the case, please answer, I am definitely willing to find a middle ground on this (it is just that I don't see it right now), as long as it keeps a high standard of quality and is not too much time consuming.

> Alternatively a setup utility script like the one used by OpenSim could be provided to help the users.

The main problem with having a setup utility that is that it is both OS dependent and environment dependent (are you using conda or not, for instance). And since free CI (such as Travis) and automatic distribution platforms (such as conda-forge) won't allow for automatically build for MATLAB, I do not have any way to test such a script extensively for C++, Python and Matlab and for Windows, Linux and Mac. Such a script would definitely breaks sooner than later and create more problems that it would solve. OpenSim are problaby using commercial CI to allow for robust installation scripting to be tested, but I cannot afford such a thing. So unless this is a deal breaker for you, I would respectfully have to reject that request...
