### 2021-02-10 pyomeca/ezc3d#177 @pariterre
@modenaxe

> the installation copied the mex files into the toolbox folder but not the dll library, which I had to copy manually in order for ezc3d to work in MATLAB. Is that the supposed behaviour?

It is now fixed in #186
Thanks for reporting!
