### 2021-02-10 f8f01d2
Added code coverage to readme
