### 2021-02-11 pyomeca/ezc3d#177 @modenaxe
@pariterre thank you for the detailed reply. My suggestions were much "lighter" in the intentions. Simply, (based on totally biased personal experience) I see the majority of MATLAB users being also Windows users and not having a broad experience with CMake, so they might just get stuck or not notice that the reason why ezc3d doesn't work is that they did not build the INSTALL solution with admin rights. If the Youtube video is hard to find (imagine no experience with GItHub either) then the risk is that they will not use the package. So I thought it would make life easier for these kind of users placing some heads up in the README. I leave this up to you - the comments above come from few years of use of BTK in biomechanical projects with engineering students.
