### 2021-02-11 #2911 @modenaxe
@Kevin-Mattheus-Moerman I am happy with both the paper and the repository (as mentioned above I checked carefully the MATLAB interface). The submission is high quality, and really solves a constant issue in biomechanics (efficiently importing gait lab data for processing)!

Thank you @pariterre, this was a great JOSS contribution to review, great fun exchanging some data structuring dilemmas as well! :smile::laughing:
