### 2021-02-17 #2911 @Kevin-Mattheus-Moerman
@pariterre I will now process this work for acceptance in JOSS.

Below are some very minor point on the paper, and some instructions for next steps:

- [ ] This sentence does not read well: `This has forced scientists to develop ad hoc solutions for each project which require to write file I/O codes for each software format files.` would the following be better?  `This has forced scientists to develop ad hoc solutions for each project which requires writing file I/O codes for each software file format.`
- [ ] Could you consider rephrasing the line `Introducing the open-source ezc3d library which provides.....`, perhaps something like these alternatives: `This work introduces....` or `Here the .... is introduced...` or `To address these issues the open source ezc3d library is introduced here....`
- [ ] I suggest replacing `open-source` by `open source`
- [ ] I suggest removal of `in-house`, since I take it all the software here is custom made as part of the open source project, it seems odd to refer to `in-house` here unless I am misunderstanding
- [ ] Instead of `dependencies-free` consider using `dependency-free`
- [ ] Since it is used once please spell out `OS` as `operational systems`
- [ ] This paper is about to be processed for acceptance, as such I recommend the authors thoroughly review it once more themselves (especially ensure the author names and acknowledgements are accurate).

- [ ] Once you have completed the above, please archive a copy of the software on [Zenodo](https://zenodo.org/) (Some users find [these steps](https://guides.github.com/activities/citable-code/) useful), and please report back here with the archive DOI.
- [ ] Ensure that the Zenodo archive metadata matches that of the paper. I.e. verify that the title and author list matches (the title might be automatically created from the repository). You may need to manually edit this information.
