### 2021-02-17 #2911 @Kevin-Mattheus-Moerman
> I think `OS` should be `operating systems` and not `operational systems`

Thanks @melund I fixed it up there :point_up:
