### 2021-02-18 fe21610
Answered editor comments and corrected typo in paper.md
