+++ doc/paper/paper.md
@@ -13,40 +13,40 @@ authors:
     orcid: 0000-0002-4107-9160
     affiliation: 1
 affiliations:
- - name: École de Kinésiologie et de Sciences de l'Activité Physique, Université de Montréal
+ - name: École de Kinésiologie et des Sciences de l'Activité Physique, Université de Montréal
    index: 1
-date: October 1st, 2020
+date: February 18th, 2021
 bibliography: paper.bib
 ---
 # Summary
-The *c3d* binary format [@crampC3dOrg2019] is an open-source standard extensively used in the field of biomechanics.
+The *c3d* binary format [@crampC3dOrg2019] is an open source standard extensively used in the field of biomechanics.
 Most of the software for biomechanical data collection and data analysis can read and export *c3d* files. 
 Initially, this format was designed to store three-dimensional point and analog data (e.g., force platform or EMG).
 Nowadays, by stretching the standard, companies have managed to integrate new types of data, including for instance IMU data.
 To match the needs of the community, Motion Lab Systems---which created and maintains the *c3d* format---regularly updates the standard to match the biomechanical needs and to include new usages of the format.
  
 Despite being extensively used by the biomechanics community, there are surprisingly few alternatives when it comes to manipulate *c3d* files outside analyses software. 
-This has forced scientists to develop *ad hoc* solutions for each project which require to write file I/O codes for each software format files. 
+This has forced scientists to develop *ad hoc* solutions for each project which requires writing file I/O codes for each software file format. 
 While it would make sense to develop a portable solution once and for all, the binary nature of the *c3d* format discourages biomechanists from digging into the trouble of developing such a solution.
 To our knowledge, `BTK` [@barreBiomechanicalToolKitBTKCore2020] is the most mature (if not, the only) biomechanics library that provides an API to read and write *c3d* files.
-Unfortunately, despite its open-sourced nature, the project has been mostly abandoned since~2016.
+Unfortunately, despite its open sourced nature, the project has been mostly abandoned since~2016.
 It gets more and more out-of-date as it does not implement the changes in the standard of the format.
 Unfortunately, due to the intricate connections of its modules, it proved difficult to update `BTK`.
  
-Introducing the open-source `ezc3d` library which provides a light and comprehensive API to easily read and write *c3d* files. 
+This work introduces the open source `ezc3d` library which provides a light and comprehensive API to easily read and write *c3d* files. 
 For lay users, `ezc3d` is an up-to-date solution to manage *c3d* files that comply with the latest recommendation of the standard.
-It also supports in-house implementations of the main biomechanics software, namely: Vicon, Qualisys, Optotrak, BTS and XSens. 
-The C++ core provides a fast file I/O library, and a convenient MATLAB and Python3 interfaces are also provided so researchers can smoothly implement `ezc3d` in their current workflow.
+It also supports *c3d* files from the main biomechanics software, namely: Vicon, Qualisys, Optotrak, BTS and XSens. 
+The C++ core provides a fast file I/O library, and convenient MATLAB and Python3 interfaces are also provided so researchers can smoothly implement `ezc3d` in their current workflow.
 In addition, since the *c3d* standard allows for multiple ways to store force platform data, a force platform analysis module is provided.
 The main feature of this module is to express forces and moments in more common reference frames---that is, expressed in the global reference frame calculated at either the origin or at the centre of pressure---so they can be directly interpreted by the user. 
  
 # The dependencies
-The `ezc3d` library was designed to be a dependencies-free library such that the lay users could easily link `ezc3d` with their project.
+The `ezc3d` library was designed to be a dependency-free library such that the lay users could easily link `ezc3d` with their project.
 Thus, by default, no dependency is needed to compile nor to use the API.
[…]
