### 2021-02-20 #2911 @Kevin-Mattheus-Moerman
@pariterre thanks for making those changes. I've ticked those boxes. One more point remains which I forgot to add to that list:

- [ ] Recently we changed [our criteria](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) to require that papers have a `Statement of need` section. Can you add such a heading in the paper and collect/formulate/move text there with a description that illustrates the research purpose of the software? I think most of this is written already so perhaps you can just collect existing text largely and put it under this heading. (note you can see examples on our website too e.g. https://joss.theoj.org/papers/10.21105/joss.02757).
