### 2021-02-20 #2911 @pariterre
@Kevin-Mattheus-Moerman

I added the `Statement of need` in https://github.com/pyomeca/ezc3d/pull/188 !
Here is the DOI for the release with the Statement of need
10.5281/zenodo.4553333

Thanks :)
