### 2021-02-20 #2911 @pariterre
Hum... I don't know where I am supposed to inform this, but the version that corresponds to the latest DOI is 1.4.2
